# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clause compliance API",
    "description": "Policy compliance service for Clause. Demo data is a fictional organization's policies, not real policy, law or regulatory guidance.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_v1_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CaseSummary_"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Create Case",
        "operationId": "create_case_api_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Case",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/messages": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Message",
        "description": "A follow-up detail. It takes effect when the next run starts.",
        "operationId": "add_message_api_v1_cases__case_id__messages_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/runs": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Start Run",
        "description": "Create a run and start it. Repeating the same Idempotency-Key returns the same run.",
        "operationId": "start_run_api_v1_cases__case_id__runs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Resume Run",
        "description": "Save clarification answers and assess again with them. Only a waiting run resumes, so\na repeated request cannot launch a second attempt.",
        "operationId": "resume_run_api_v1_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Cancel Run",
        "description": "Stop further stages. A stage already waiting on the model may still finish, but its\nresult is never published. Canceling a finished run changes nothing.",
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Run Events",
        "description": "Server-sent events. Replays events after `after` (or the Last-Event-ID a reconnecting\nbrowser sends), then streams new ones until the run ends. A 204 tells the browser the run\nhas ended and there is nothing more, so it stops reconnecting.",
        "operationId": "run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "last-event-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "204": {
            "description": "Run ended"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssessmentStatus": {
        "type": "string",
        "enum": [
          "non_compliant",
          "conflicting_policy",
          "insufficient_information",
          "compliant_within_scope",
          "out_of_scope"
        ],
        "title": "AssessmentStatus",
        "description": "Displayed result, chosen by the ordered rules in §5.2 (first match wins)."
      },
      "CaseDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "business_area": {
            "type": "string",
            "title": "Business Area"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunState"
              },
              {
                "type": "null"
              }
            ]
          },
          "unresolved_facts": {
            "type": "integer",
            "title": "Unresolved Facts"
          },
          "review_state": {
            "$ref": "#/components/schemas/ReviewState"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "scenario_text": {
            "type": "string",
            "title": "Scenario Text"
          },
          "policy_snapshot_id": {
            "type": "string",
            "title": "Policy Snapshot Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Facts"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Assessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            },
            "type": "array",
            "title": "Pending Questions"
          },
          "agent_messages": {
            "items": {
              "$ref": "#/components/schemas/AgentMessage"
            },
            "type": "array",
            "title": "Agent Messages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "owner",
          "business_area",
          "status",
          "run_state",
          "unresolved_facts",
          "review_state",
          "updated_at",
          "as_of",
          "scope",
          "scenario_text",
          "policy_snapshot_id",
          "messages",
          "facts",
          "latest_run_id",
          "assessment",
          "pending_questions",
          "agent_messages"
        ],
        "title": "CaseDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "role",
          "text",
          "created_at"
        ],
        "title": "Message"
      },
      "NewCaseRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 20,
            "title": "Text"
          },
          "business_area": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Business Area"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "business_area",
          "as_of"
        ],
        "title": "NewCaseRequest"
      },
      "NewMessageRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "NewMessageRequest"
      },
...
      "Page_CaseSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[CaseSummary]"
      },
...
      "ResumeRequest": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 4000
                },
                {
                  "type": "null"
                }
              ]
            },
            "propertyNames": {
              "maxLength": 128,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 3,
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "ResumeRequest",
        "description": "Clarification answers keyed by question ID; `null` means \"I don't know\"."
      },
      "ReviewState": {
        "type": "string",
        "enum": [
          "unreviewed",
          "accepted",
          "challenged",
          "information_requested"
        ],
        "title": "ReviewState"
      },
...
      "RunState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "waiting_for_user",
          "completed",
          "failed",
          "canceled"
        ],
        "title": "RunState",
        "description": "Workflow state. A timeout is `failed`, never a compliance verdict."
      },
...
    }
  }
}